moved tooltip to event renderer split up event settings into two dialogs
